Add option to configure package only, handle long counter names by using sub directories

+++ src/main/java/org/apache/sling/metrics/impl/dropwizard/CsvReporter.java
@@ -1,9 +1,5 @@
 package org.apache.sling.metrics.impl.dropwizard;
 
-import com.codahale.metrics.Clock;
-import com.codahale.metrics.MetricFilter;
-import com.codahale.metrics.MetricRegistry;
-
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -11,24 +7,26 @@ import java.io.OutputStreamWriter;
 import java.io.PrintWriter;
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
+import java.nio.charset.Charset;
+import java.util.Iterator;
 import java.util.Locale;
+import java.util.Map.Entry;
+import java.util.SortedMap;
 import java.util.concurrent.TimeUnit;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import com.codahale.metrics.Clock;
 import com.codahale.metrics.Counter;
 import com.codahale.metrics.Gauge;
 import com.codahale.metrics.Histogram;
 import com.codahale.metrics.Meter;
+import com.codahale.metrics.MetricFilter;
+import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.ScheduledReporter;
 import com.codahale.metrics.Snapshot;
 import com.codahale.metrics.Timer;
-
-import java.nio.charset.Charset;
-import java.util.Iterator;
-import java.util.SortedMap;
-import java.util.Map.Entry;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 // This class came from Drowizard 3.1. The sanitise method was broken, fixed here.
 
@@ -50,6 +48,7 @@ public class CsvReporter extends ScheduledReporter {
         this.clock = clock;
     }
 
+    @Override
     public void report(SortedMap<String, Gauge> gauges, SortedMap<String, Counter> counters, SortedMap<String, Histogram> histograms, SortedMap<String, Meter> meters, SortedMap<String, Timer> timers) {
         long timestamp = TimeUnit.MILLISECONDS.toSeconds(this.clock.getTime());
         Iterator var8 = gauges.entrySet().iterator();
@@ -92,7 +91,7 @@ public class CsvReporter extends ScheduledReporter {
 
     private void reportTimer(long timestamp, String name, Timer timer) {
         Snapshot snapshot = timer.getSnapshot();
-        this.report(timestamp, name, "count,max,mean,min,stddev,p50,p75,p95,p98,p99,p999,mean_rate,m1_rate,m5_rate,m15_rate,rate_unit,duration_unit", "%d,%f,%f,%f,%f,%f,%f,%f,%f,%f,%f,%f,%f,%f,%f,calls/%s,%s", new Object[]{Long.valueOf(timer.getCount()), Double.valueOf(this.convertDuration((double)snapshot.getMax())), Double.valueOf(this.convertDuration(snapshot.getMean())), Double.valueOf(this.convertDuration((double)snapshot.getMin())), Double.valueOf(this.convertDuration(snapshot.getStdDev())), Double.valueOf(this.convertDuration(snapshot.getMedian())), Double.valueOf(this.convertDuration(snapshot.get75thPercentile())), Double.valueOf(this.convertDuration(snapshot.get95thPercentile())), Double.valueOf(this.convertDuration(snapshot.get98thPercentile())), Double.valueOf(this.convertDuration(snapshot.get99thPercentile())), Double.valueOf(this.convertDuration(snapshot.get999thPercentile())), Double.valueOf(this.convertRate(timer.getMeanRate())), Double.valueOf(this.convertRate(timer.getOneMinuteRate())), Double.valueOf(this.convertRate(timer.getFiveMinuteRate())), Double.valueOf(this.convertRate(timer.getFifteenMinuteRate())), this.getRateUnit(), this.getDurationUnit()});
+        this.report(timestamp, name, "count,max,mean,min,stddev,p50,p75,p95,p98,p99,p999,mean_rate,m1_rate,m5_rate,m15_rate,rate_unit,duration_unit", "%d,%f,%f,%f,%f,%f,%f,%f,%f,%f,%f,%f,%f,%f,%f,calls/%s,%s", new Object[]{Long.valueOf(timer.getCount()), Double.valueOf(this.convertDuration(snapshot.getMax())), Double.valueOf(this.convertDuration(snapshot.getMean())), Double.valueOf(this.convertDuration(snapshot.getMin())), Double.valueOf(this.convertDuration(snapshot.getStdDev())), Double.valueOf(this.convertDuration(snapshot.getMedian())), Double.valueOf(this.convertDuration(snapshot.get75thPercentile())), Double.valueOf(this.convertDuration(snapshot.get95thPercentile())), Double.valueOf(this.convertDuration(snapshot.get98thPercentile())), Double.valueOf(this.convertDuration(snapshot.get99thPercentile())), Double.valueOf(this.convertDuration(snapshot.get999thPercentile())), Double.valueOf(this.convertRate(timer.getMeanRate())), Double.valueOf(this.convertRate(timer.getOneMinuteRate())), Double.valueOf(this.convertRate(timer.getFiveMinuteRate())), Double.valueOf(this.convertRate(timer.getFifteenMinuteRate())), this.getRateUnit(), this.getDurationUnit()});
     }
 
     private void reportMeter(long timestamp, String name, Meter meter) {
@@ -116,7 +115,7 @@ public class CsvReporter extends ScheduledReporter {
         try {
             File e = new File(this.directory, this.sanitize(name) + ".csv");
             boolean fileAlreadyExists = e.exists();
-            if(fileAlreadyExists || e.createNewFile()) {
+            if(fileAlreadyExists || (e.getParentFile().mkdirs() && e.createNewFile())) {
                 PrintWriter out = new PrintWriter(new OutputStreamWriter(new FileOutputStream(e, true), UTF_8));
 
                 try {
@@ -137,7 +136,7 @@ public class CsvReporter extends ScheduledReporter {
 
     protected String sanitize(String name) {
         try {
-            return URLEncoder.encode(name, "UTF-8");
+            return URLEncoder.encode(name, "UTF-8").replace('*', File.separatorChar);
         } catch (UnsupportedEncodingException e) {
             throw new RuntimeException(e);
         }
